requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai==0.28.0
gunicorn==21.2.0
requests==2.31.0
jinja2==3.1.3
werkzeug==3.0.1
itsdangerous==2.1.2
urllib3==1.26.18 